Move from_state/2 to Membrane.Core.*
