+++ lib/membrane/core/component.ex
@@ -7,9 +7,9 @@ defmodule Membrane.Core.Component do
           | Membrane.Core.Element.State.t()
 
   @type callback_context_option_t ::
-          Membrane.Element.CallbackContext.option_t()
-          | Membrane.Bin.CallbackContext.option_t()
-          | Membrane.Pipeline.CallbackContext.option_t()
+          Membrane.Core.Element.CallbackContext.option_t()
+          | Membrane.Core.Bin.CallbackContext.option_t()
+          | Membrane.Core.Pipeline.CallbackContext.option_t()
 
   @type callback_context_options_t :: [callback_context_option_t()]
 
@@ -27,61 +27,18 @@ defmodule Membrane.Core.Component do
 
   @spec callback_context(state_t(), callback_context_options_t()) ::
           callback_context_t()
-  # def callback_context(state, args \\ [])
-
-  # [Pipeline, Bin, Element]
-  # |> Enum.map(fn component ->
-  #   def callback_context(
-  #         %unquote(Module.concat([Membrane.Core, component, State])){} = state,
-  #         args
-  #       ),
-  #       do: unquote(Module.concat([Membrane, component, CallbackContext])).from_state(state, args)
-  # end)
-
   def callback_context(state, args \\ []) do
+    alias Membrane.Core.{Bin, Element, Pipeline}
+
     callback_context_module =
       case state do
-        %Membrane.Core.Element.State{} -> Membrane.Element.CallbackContext
-        %Membrane.Core.Bin.State{} -> Membrane.Bin.CallbackContext
-        %Membrane.Core.Pipeline.State{} -> Membrane.Pipeline.CallbackContext
+        %Element.State{} -> Element.CallbackContext
+        %Bin.State{} -> Bin.CallbackContext
+        %Pipeline.State{} -> Pipeline.CallbackContext
       end
 
     callback_context_module.from_state(state, args)
   end
-
-  # defmacro callback_context_generator(restrict, module, state, args \\ []) do
-  #   module = Macro.expand(module, __ENV__)
-
-  #   restrict =
-  #     case restrict do
-  #       :parent -> [Pipeline, Bin]
-  #       :child -> [Bin, Element]
-  #       :any -> [Pipeline, Bin, Element]
-  #       restrict -> restrict
-  #     end
-
-  #   requires =
-  #     restrict
-  #     |> Enum.map(fn component ->
-  #       quote do
-  #         require unquote(context(component, module))
-  #       end
-  #     end)
-
-  #   clauses =
-  #     restrict
-  #     |> Enum.flat_map(fn component ->
-  #       quote do
-  #         %unquote(state(component)){} ->
-  #           &unquote(context(component, module)).from_state(&1, unquote(args))
-  #       end
-  #     end)
-
-  #   quote do
-  #     unquote_splicing(requires)
-  #     unquote({:case, [], [state, [do: clauses]]})
-  #   end
-  # end
 
   @spec is_pipeline?(state_t) :: boolean()
   def is_pipeline?(%Membrane.Core.Pipeline.State{}), do: true
@@ -100,9 +57,4 @@ defmodule Membrane.Core.Component do
 
   @spec is_parent?(state_t) :: boolean()
   def is_parent?(state), do: is_pipeline?(state) or is_bin?(state)
-
-  # defp context(component, module),
-  #   do: Module.concat([Membrane, component, CallbackContext, module])
-
-  # defp state(component), do: Module.concat([Membrane.Core, component, State])
 end
