+++ lib/membrane/core/element/demand_controller.ex
@@ -7,8 +7,7 @@ defmodule Membrane.Core.Element.DemandController do
 
   alias Membrane.Core.{CallbackHandler, Message}
   alias Membrane.Core.Child.PadModel
-  alias Membrane.Core.Element.{ActionHandler, PlaybackQueue, State, Toilet}
-  alias Membrane.Element.CallbackContext
+  alias Membrane.Core.Element.{ActionHandler, CallbackContext, PlaybackQueue, State, Toilet}
   alias Membrane.Pad
 
   require Membrane.Core.Child.PadModel
