+++ lib/membrane/core/element/lifecycle_controller.ex
@@ -8,8 +8,7 @@ defmodule Membrane.Core.Element.LifecycleController do
 
   alias Membrane.{Clock, Element, Sync}
   alias Membrane.Core.{CallbackHandler, Child, Element, Message}
-  alias Membrane.Core.Element.{ActionHandler, PlaybackQueue, State}
-  alias Membrane.Element.CallbackContext
+  alias Membrane.Core.Element.{ActionHandler, CallbackContext, PlaybackQueue, State}
 
   require Membrane.Core.Child.PadModel
   require Membrane.Core.Message
