+++ lib/membrane/core/element/pad_controller.ex
@@ -10,6 +10,7 @@ defmodule Membrane.Core.Element.PadController do
 
   alias Membrane.Core.Element.{
     ActionHandler,
+    CallbackContext,
     DemandController,
     EventController,
     InputQueue,
@@ -19,11 +20,9 @@ defmodule Membrane.Core.Element.PadController do
   }
 
   alias Membrane.Core.Parent.Link.Endpoint
-  alias Membrane.Element.CallbackContext
 
   require Membrane.Core.Child.PadModel
   require Membrane.Core.Message
-  require Membrane.Element.CallbackContext
   require Membrane.Logger
   require Membrane.Pad
