+++ lib/membrane/element/callback_context.ex
@@ -1,5 +1,7 @@
 defmodule Membrane.Element.CallbackContext do
-  @moduledoc false
+  @moduledoc """
+  Describes context passed to Membrane Elements callbacks.
+  """
 
   @type t :: %{
           :pads => %{Membrane.Pad.ref_t() => Membrane.Element.PadData.t()},
@@ -13,25 +15,4 @@ defmodule Membrane.Element.CallbackContext do
           optional(:pad_options) => map(),
           optional(:old_stream_format) => Membrane.StreamFormat.t()
         }
-
-  @type option_t ::
-          {:incoming_demand, non_neg_integer()}
-          | {:pad_options, map()}
-          | {:old_stream_format, Membrane.StreamFormat.t()}
-
-  @type options_t :: [option_t()]
-
-  @spec from_state(Membrane.Core.Element.State.t(), options_t()) :: t()
-  def from_state(state, additional_fields \\ []) do
-    Map.new(additional_fields)
-    |> Map.merge(%{
-      pads: state.pads_data,
-      clock: state.synchronization.clock,
-      parent_clock: state.synchronization.parent_clock,
-      name: state.name,
-      playback: state.playback,
-      resource_guard: state.resource_guard,
-      utility_supervisor: state.subprocess_supervisor
-    })
-  end
 end
